Update stop.sh
stop.sh脚本目前，只杀死collector.sh启动的进程，并未杀死hypervisor的守护进程，stop后collector进程又会自动拉起，并未达到停止服务的目的。
修改后，可直接停止collector和hypervisor两个进程，彻底停掉mongo-shake服务运行。

+++ scripts/stop.sh
@@ -1,2 +1,13 @@
 #!/usr/bin/env bash
-kill -9 "$(cat "$1")"
+#kill -9 "$(cat "$1")"
+if [ $# != 1 ] ; then
+    echo "USAGE: $0 [pidfile]"
+    exit 0
+fi
+ppid=$(ps -ef | grep -v grep |awk '{if ($2=='`cat $1`') print $3}')
+[ -z $ppid ] && echo "[Fail] No process number for $(cat "$1")." && exit 1
+if [ $ppid -eq 1 ];then
+    kill -9 "$(cat "$1")"
+else
+    kill -9 "$(cat "$1")" "$ppid"
+fi
